moved section/segment code to sections.go. Fixed some warnings about comments
